added more components

+++ src/app/api/salons/[id]/route.ts
@@ -1,19 +1,6 @@
 import { NextResponse } from "next/server";
 import { prisma } from "../../../../../prisma/prisma";
 
-export async function GET(
-  req: Request,
-  { params }: { params: { id: string } }
-) {
-  const salon = await prisma.salon.findUnique({ where: { id: params.id } });
-  if (!salon)
-    return NextResponse.json(
-      { error: "Salão não encontrado." },
-      { status: 404 }
-    );
-  return NextResponse.json(salon);
-}
-
 export async function PUT(
   req: Request,
   { params }: { params: { id: string } }
@@ -27,6 +14,24 @@ export async function DELETE(
   req: Request,
   { params }: { params: { id: string } }
 ) {
-  await prisma.salon.delete({ where: { id: params.id } });
-  return NextResponse.json({ message: "Salão deletado." });
+  try {
+    const { id } = await params;
+    const salon = await prisma.salon.findUnique({
+      where: { id },
+    });
+    if (!salon) {
+      return NextResponse.json(
+        { error: "Salão não encontrado." },
+        { status: 404 }
+      );
+    }
+    await prisma.salon.delete({ where: { id: params.id } });
+
+    return NextResponse.json({ message: "Salão deletado com sucesso." });
+  } catch (error) {
+    return NextResponse.json(
+      { error: "Ocorreu um erro ao deletar o salão." },
+      { status: 500 }
+    );
+  }
 }
